# docs/api-contract-project.yaml
openapi: 3.0.0
info:
  title: API Contract - Sistem Penulisan Hibah Buku (Full Project)
  description: |
    API Master Contract for all 4 modules in the Book Grant Writing System.
    (Kontrak Master API untuk ke-4 modul pada Sistem Penulisan Hibah Buku.)

    Termasuk: Modul 1 Fondasi Sistem, User, Role & Kontrak (Kelompok 4)
    Modul 2 Penulis & Manajemen Naskah (Kelompok 3)
    Modul 3 Reviewer & Proses Review (Kelompok 2)
    Modul 4 Penerbit, Pra-Cetak & Monitoring (Kelompok 1)
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local Development Server / Server Pengembangan Lokal
...
paths:
...
  /auth/login:
    post:
      summary: Login Pengguna
      description: |
        Endpoint login untuk semua role (admin, penulis, reviewer, penerbit).
        Mengembalikan Bearer Token untuk autentikasi request selanjutnya.
      tags:
        - Modul 1 - Fondasi Sistem & Kontrak (Kelompok 4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@hibahbuku.id"
                password:
                  type: string
                  format: password
                  example: "admin123"
      responses:
        "200":
          description: Login berhasil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login berhasil."
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      user:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          email:
                            type: string
                          role:
                            type: string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/logout:
    post:
      summary: Logout Pengguna
      tags:
        - Modul 1 - Fondasi Sistem & Kontrak (Kelompok 4)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout berhasil.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /auth/me:
    get:
      summary: Data User yang Sedang Login
      tags:
        - Modul 1 - Fondasi Sistem & Kontrak (Kelompok 4)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Data user berhasil diambil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /admin/users:
    get:
      summary: Daftar Semua User
      tags:
        - Modul 1 - Fondasi Sistem & Kontrak (Kelompok 4)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [penulis, reviewer, penerbit]
      responses:
        "200":
          description: Daftar user berhasil diambil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
    post:
      summary: Buat Akun Reviewer atau Penerbit
      tags:
        - Modul 1 - Fondasi Sistem & Kontrak (Kelompok 4)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, role, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                role:
                  type: string
                  enum: [reviewer, penerbit]
                password:
                  type: string
      responses:
        "201":
          description: Akun berhasil dibuat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponseObject"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /author-confirmations:
    post:
      summary: Submit Form Kesediaan Penulis
      description: Endpoint publik (tanpa login).
      tags:
        - Modul 1 - Fondasi Sistem & Kontrak (Kelompok 4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, book_title, book_type]
              properties:
                name:
                  type: string
                email:
                  type: string
                institution:
                  type: string
                phone:
                  type: string
                book_title:
                  type: string
                book_type:
                  type: string
                  enum: [ajar, referensi]
      responses:
        "201":
          description: Form kesediaan berhasil disubmit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      summary: Daftar Form Kesediaan (Admin)
      tags:
        - Modul 1 - Fondasi Sistem & Kontrak (Kelompok 4)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processed]
      responses:
        "200":
          description: Daftar form kesediaan berhasil diambil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuthorConfirmation"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ==========================================
  # MODUL 2: PENULIS & MANAJEMEN NASKAH (KLP 3)
  # ==========================================
...
  /publisher/dashboard:
    get:
      summary: Publisher Dashboard (Dasbor Penerbit)
      description: |
        Menampilkan ringkasan tugas penerbit:
        - jumlah naskah pra-cetak (status 'preprint')
        - jumlah approved bulan ini (dari publisher_checks.decision)
        - jumlah revised bulan ini
        - daftar 5 naskah pra-cetak terbaru dengan tautan
      tags:
        - Modul 4 - Penerbit, Pra-Cetak, & Monitoring (Kelompok 1)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Data dashboard penerbit berhasil diambil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      preprint_count:
                        type: integer
                      approved_count_this_month:
                        type: integer
                      revised_count_this_month:
                        type: integer
                      latest_preprints:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            title:
                              type: string
                            author:
                              type: string
                            submitted_at:
                              type: string
                              format: date-time
                            _links:
                              type: object
                              properties:
                                check_detail:
                                  type: object
                                  properties:
                                    href:
                                      type: string
                                      example: "/api/publisher/check/{id}"
                                    method:
                                      type: string
                                      example: GET
                  _links:
                    type: object
                    properties:
                      preprint_list:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/publisher/preprint-list"
                          method:
                            type: string
                            example: GET
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
  /publisher/check/{manuscriptId}:
    get:
      summary: Get Preprint Check Data
      description: |
        Mengembalikan data naskah (dari manuscripts, users, manuscript_files) 
        dan hasil pemeriksaan (dari publisher_checks) jika sudah ada.
      tags:
        - Modul 4 - Penerbit, Pra-Cetak, & Monitoring (Kelompok 1)
      security:
        - bearerAuth: []
      parameters:
        - name: manuscriptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Data pemeriksaan berhasil diambil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      manuscript:
                        type: object
                        properties:
                          id:
                            type: integer
                          title:
                            type: string
                          author_name:
                            type: string
                          cover_file_url:
                            type: string
                          page_count:
                            type: integer
                          admin_documents:
                            type: array
                            items:
                              type: object
                      check_result:
                        type: object
                        nullable: true
                        properties:
                          cover_ok:
                            type: boolean
                          pages_ok:
                            type: boolean
                          admin_ok:
                            type: boolean
                          notes:
                            type: string
                          decision:
                            type: string
                            enum: [approved, revised]
                            nullable: true
                  _links:
                    type: object
                    properties:
                      save_check:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/publisher/check/{manuscriptId}"
                          method:
                            type: string
                            example: POST
                      make_decision:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/publisher/decision"
                          method:
                            type: string
                            example: POST
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      summary: Save Preprint Check
      description: |
        Menyimpan hasil pemeriksaan kelengkapan (cover, halaman, dokumen admin) 
        ke tabel `publisher_checks`. Jika sudah ada, akan diupdate.
      tags:
        - Modul 4 - Penerbit, Pra-Cetak, & Monitoring (Kelompok 1)
      security:
        - bearerAuth: []
      parameters:
        - name: manuscriptId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cover_ok
                - pages_ok
                - admin_ok
              properties:
                cover_ok:
                  type: boolean
                pages_ok:
                  type: boolean
                admin_ok:
                  type: boolean
                notes:
                  type: string
      responses:
        "200":
          description: Pemeriksaan berhasil disimpan.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  _links:
                    type: object
                    properties:
                      check_detail:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/publisher/check/{manuscriptId}"
                          method:
                            type: string
                            example: GET
                      make_decision:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/publisher/decision"
                          method:
                            type: string
                            example: POST
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /publisher/decision:
    post:
      summary: Publisher Decision
      description: |
        Memberikan keputusan akhir penerbit (approved / revised).
        Data keputusan (decision) disimpan di tabel `publisher_checks`.
        Jika keputusan adalah 'revised', catatan revisi (revision_notes) akan dikirim ke Modul 2 (Manajemen Naskah) untuk ditampilkan kepada penulis.
        Wajib menyertakan `check_id` (foreign key ke publisher_checks) yang diperoleh dari pemeriksaan sebelumnya.
      tags:
        - Modul 4 - Penerbit, Pra-Cetak, & Monitoring (Kelompok 1)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - check_id
                - decision
              properties:
                check_id:
                  type: integer
                  description: ID dari record publisher_checks yang terkait.
                decision:
                  type: string
                  enum: [approved, revised]
                revision_notes:
                  type: string
                  description: Catatan revisi (wajib diisi jika decision = revised). Akan diteruskan ke Modul 2.
      responses:
        "200":
          description: Keputusan berhasil disimpan.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      decision_id:
                        type: integer
                        description: ID dari publisher_checks yang diupdate.
                      status:
                        type: string
                  _links:
                    type: object
                    properties:
                      dashboard:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/publisher/dashboard"
                          method:
                            type: string
                            example: GET
                      preprint_list:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/publisher/preprint-list"
                          method:
                            type: string
                            example: GET
        "400":
          description: Bad request (misal revision_notes tidak diisi)
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /notification/send:
    post:
      summary: Send Email Notification
      description: |
        Endpoint internal untuk mengirim email.
        Request berisi `type` (event_type) yang akan dicatat di `notification_log`.
        Setelah berhasil, log disimpan di `notification_log`.
      tags:
        - Modul 4 - Penerbit, Pra-Cetak, & Monitoring (Kelompok 1)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - subject
                - body
                - type
              properties:
                to:
                  type: string
                  format: email
                subject:
                  type: string
                body:
                  type: string
                type:
                  type: string
                  enum:
                    [
                      account_created,
                      contract_validated,
                      draft_uploaded,
                      review_assigned,
                      review_completed,
                      revision_requested,
                      preprint_entered,
                      publisher_approved,
                      publisher_revised,
                      deadline_reminder,
                    ]
      responses:
        "200":
          description: Email berhasil dikirim.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  _links:
                    type: object
                    properties:
                      notification_logs:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/admin/notification-logs"
                          method:
                            type: string
                            example: GET
        "400":
          description: Data tidak valid
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /reminder/trigger:
    post:
      summary: Trigger Reminder Manually
      description: |
        Menjalankan cron reminder secara manual.
        Membaca tabel `deadlines` untuk mengirim pengingat.
        (Catatan: Skema SQL tidak menyediakan tabel reminder_logs; reminder tetap dijalankan berdasarkan deadlines.)
      tags:
        - Modul 4 - Penerbit, Pra-Cetak, & Monitoring (Kelompok 1)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reminder berhasil diproses.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      reminders_sent:
                        type: integer
                      reminders_failed:
                        type: integer
                  _links:
                    type: object
                    properties:
                      dashboard:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/admin/dashboard-stats"
                          method:
                            type: string
                            example: GET
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message.
          example: "Operation successful."
        data:
          nullable: true
    SuccessResponseObject:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Data retrieved successfully."
        data:
          type: object
...
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Data tidak valid."
        data:
          type: object
          properties:
            errors:
              type: object
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
    AuthorConfirmation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        institution:
          type: string
        phone:
          type: string
        book_title:
          type: string
        book_type:
          type: string
        status:
          type: string
        created_at:
          type: string
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Unauthenticated."
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Access denied."
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Resource not found."
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'